+++ pkg/api/validation/dynakube/preview.go
@@ -1,0 +1,11 @@
+// Copyright Dynatrace LLC
+// SPDX-License-Identifier: Apache-2.0
+
+package validation
+
+const (
+	featurePreviewWarningMessage = `%s feature is in PREVIEW.`                                            //nolint:unused
+	basePreviewWarning           = "PREVIEW features are NOT production ready and you may run into bugs." //nolint:unused
+)
+
+// Logics related to preview warnings go here.
